fix(cli): round-1 review feedback (codex + claude + coderabbit)
Multi-party review of #585 found 6 real bugs and several robustness
gaps. Fixes:

Must-fix
- login: --password - read via io.ReadAll instead of a single 4096-byte
  Read(), so piped passwords don't get silently truncated. Capped at
  1 MiB to fail fast on misuse.
- user update: refactor to *string + cmd.Flags().Changed() semantics
  (consistent with token/invite/route updates); rejects "no fields"
  invocation instead of silently PUT-ing {}.
- user create / user password: --dry-run no longer triggers an
  interactive prompt — preview prints "<redacted>" without ever reading
  the password.
- strategy update: -f no longer defaults to "-" (would block on a TTY
  if the flag was forgotten); requires an explicit path or '-'.
- provider import: strip UTF-8 BOM before sniffing the first byte, so
  Notepad-saved JSON imports cleanly.
- output.Truncate: count runes, not bytes — was splitting multi-byte
  UTF-8 sequences mid-character.

Issue
- api.NewFromContext: validate URL scheme (http|https) and non-empty
  host so "localhost:9880" no longer reaches the wire as a path.
- api.extractErrorMsg: also surface "message" and "detail" fields,
  not just "error", so non-maxx middleware errors aren't dropped.
- api: cap response body at 32 MiB via io.LimitReader.
- api: print a per-invocation warning when a context has
  InsecureSkipVerify=true.
- cfg.Load: surface tryMigrate errors instead of swallowing them.
- cfg.saveTo: use os.CreateTemp with a random tmp name + same-dir
  rename, avoiding TOCTOU/symlink concerns and concurrent-save
  collisions; explicit cleanup on failure.
- provider export: Use line is now "[--out file.json]" (was "-o" which
  collided with the root output-format flag).
- route update: cast --project-id / --provider-id / --retry-config-id
  to uint64 to match the server's domain type.

Nits
- topics.go: comment claimed Hidden:false hid topics from completion;
  rewrote to explain they are intentionally visible (so AI agents
  discover "help reference").
- output.go: drop misleading "stderr-style" wording.

Tests
- output: Truncate(ASCII/UTF-8 multibyte/zero/negative).
- api: NewFromContext rejects bare host:port and unknown schemes,
  accepts http/https; extractErrorMsg picks error|message|detail.
- cmd: provider import strips UTF-8 BOM.
- cmd: every root subcommand must have a known GroupID (catches the
  next "forgot to group it" mistake).

+++ cmd/maxx-cli/cmd/route.go
@@ -170,7 +170,7 @@ instead, which overrides any individual flags.`,
 					if projectID < 0 {
 						return fmt.Errorf("--project-id must be >= 0")
 					}
-					patch["projectID"] = projectID
+					patch["projectID"] = uint64(projectID)
 				}
 				if cmd.Flags().Changed("client-type") {
 					patch["clientType"] = clientType
@@ -179,7 +179,7 @@ instead, which overrides any individual flags.`,
 					if providerID < 0 {
 						return fmt.Errorf("--provider-id must be >= 0")
 					}
-					patch["providerID"] = providerID
+					patch["providerID"] = uint64(providerID)
 				}
 				if cmd.Flags().Changed("position") {
 					patch["position"] = position
@@ -194,7 +194,7 @@ instead, which overrides any individual flags.`,
 					if retryConfigID < 0 {
 						return fmt.Errorf("--retry-config-id must be >= 0")
 					}
-					patch["retryConfigID"] = retryConfigID
+					patch["retryConfigID"] = uint64(retryConfigID)
 				}
 				if len(patch) == 0 {
 					return fmt.Errorf("no fields to update; pass one of --enabled/--native/--project-id/--client-type/--provider-id/--position/--weight/--retry-config-id, or use -f")
